move data loader back into experiment, strore metadata
