+++ mdsuite/experiment/experiment.py
@@ -33,15 +33,28 @@ from mdsuite.utils.units import units_dict, Units
 from mdsuite.database.experiment_database import ExperimentDatabase
 from mdsuite.visualizer.trajectory_visualizer import SimulationVisualizer
 
-from .add_files import ExperimentAddingFiles
 from .run_module import RunModule
 
 from typing import Union
 
 log = logging.getLogger(__name__)
 
+from mdsuite.database.simulation_database import Database
+from mdsuite.file_io.file_io_dict import dict_file_io
+from mdsuite.file_io.file_read import FileProcessor
+from mdsuite.utils.exceptions import ElementMassAssignedZero
+from mdsuite.utils.meta_functions import join_path
+import mdsuite.utils.meta_functions
 
-class Experiment(ExperimentDatabase, ExperimentAddingFiles):
+import pathlib
+import pubchempy as pcp
+import importlib.resources
+import json
+import numpy as np
+import copy
+
+
+class Experiment(ExperimentDatabase):
     """
     The central experiment class fundamental to all analysis.
 
@@ -72,13 +85,13 @@ class Experiment(ExperimentDatabase, ExperimentAddingFiles):
     """
 
     def __init__(
-        self,
-        project,
-        experiment_name,
-        time_step=None,
-        temperature=None,
-        units: Union[str, Units] = None,
-        cluster_mode=False,
+            self,
+            project,
+            experiment_name,
+            time_step=None,
+            temperature=None,
+            units: Union[str, Units] = None,
+            cluster_mode=False,
     ):
         """
         Initialise the experiment class.
@@ -135,7 +148,6 @@ class Experiment(ExperimentDatabase, ExperimentAddingFiles):
         self.sample_rate = (
             None  # Rate at which configurations are dumped in the trajectory.
         )
-        self.batch_size = None  # Number of configurations in each batch.
         self.volume = None
         self.properties = None  # Properties measured in the simulation.
         self.property_groups = (
@@ -294,11 +306,11 @@ class Experiment(ExperimentDatabase, ExperimentAddingFiles):
         transformation_run.run_transformation()  # perform the transformation
 
     def run_visualization(
-        self,
-        species: list = None,
-        molecules: bool = False,
-        unwrapped: bool = False,
-        starting_configuration: int = None,
+            self,
+            species: list = None,
+            molecules: bool = False,
+            unwrapped: bool = False,
+            starting_configuration: int = None,
     ):
         """
         Perform a trajectory visualization.
@@ -404,3 +416,218 @@ class Experiment(ExperimentDatabase, ExperimentAddingFiles):
         species = self.species
         species[element]["mass"] = mass  # update the mass
         self.species = species
+
+    def add_data(
+            self,
+            trajectory_file: str = None,
+            file_format: str = "lammps_traj",
+            force: bool = False,
+            update_with_pubchempy=True
+    ):
+        """
+        Add tensor_values to the database_path
+
+        Parameters
+        ----------
+        file_format :
+                Format of the file being read in. Default is file_path
+        trajectory_file : str
+                Trajectory file to be process and added to the database_path.
+        force : bool
+                If true, a file will be read regardless of if it has already
+                been seen.
+        update_with_pubchempy: bool
+                Whether or not to look for the masses of the species in pubchempy
+        """
+
+        # Check if there is a trajectory file.
+        if trajectory_file is None:
+            raise ValueError("You need to pass a trajectory_file for data loading")
+
+        # make absolute path
+        traj_file_path = pathlib.Path(trajectory_file)
+        traj_file_path.resolve()
+
+        already_read = traj_file_path in self.read_files
+        if already_read and not force:
+            log.info(
+                "This file has already been read, skipping this now."
+                "If this is not desired, please add force=True "
+                "to the command."
+            )
+            return  # End the method.
+
+        # todo trajectory_reader should be an argument of this function
+        trajectory_reader, file_type = _load_trajectory_reader(
+            file_format, trajectory_file
+        )
+        database = Database(
+            name=pathlib.Path(self.database_path, "database.hdf5").as_posix(),
+        )
+
+        # Check to see if a database_path exists
+        # todo make method of database class
+        database_path = pathlib.Path(
+            self.database_path, "database.hdf5"
+        )  # get theoretical path.
+
+        metadata = trajectory_reader.get_metadata()
+        architecture = _species_list_to_architecture_dict(metadata.species_list,
+                                                          metadata.n_configurations)
+        if not database_path.exists():
+            self._store_metadata(metadata, update_with_pubchempy=update_with_pubchempy)
+            database.initialize_database(architecture)
+        else:
+            database.resize_datasets(architecture)
+
+        for i, batch in enumerate(trajectory_reader.get_configurations_generator()):
+            database.add_data(
+                chunk=batch,
+                start_idx=self.number_of_configurations
+            )
+            self.number_of_configurations += batch.chunk_size
+
+        self.version += 1
+
+        self.memory_requirements = database.get_memory_information()
+
+        # set at the end, because if something fails, the file was not properly read.
+        self.read_files = self.read_files + [traj_file_path]
+
+    def load_matrix(
+            self,
+            property_name: str = None,
+            species: list = None,
+            select_slice: np.s_ = None,
+            path: list = None,
+    ):
+        """
+        Load a desired property matrix.
+
+        Parameters
+        ----------
+        property_name : str
+                Name of the matrix to be loaded, e.g. 'Unwrapped_Positions', 'Velocities'
+        species : list
+                List of species to be loaded
+        select_slice : np.slice
+                A slice to select from the database_path.
+        path : str
+                optional path to the database_path.
+
+        Returns
+        -------
+        property_matrix : np.array, tf.tensor
+                Tensor of the property to be studied. Format depends on kwargs.
+        """
+        database = Database(name=pathlib.Path(self.database_path, "database.hdf5").as_posix())
+
+        if path is not None:
+            return database.load_data(path_list=path, select_slice=select_slice)
+
+        else:
+            # If no species list is given, use all species in the Experiment.
+            if species is None:
+                species = list(self.species.keys())
+            # If no slice is given, load all configurations.
+            if select_slice is None:
+                select_slice = np.s_[:]  # set the numpy slice object.
+
+        path_list = []
+        for item in species:
+            path_list.append(join_path(item, property_name))
+        return database.load_data(path_list=path_list, select_slice=select_slice)
+
+    def _store_metadata(self, metadata, update_with_pubchempy = False):
+        # new trajectory: store all metadata and construct a new database
+        self.temperature = metadata.temperature
+        self.box_array = metadata.box_l
+        self.dimensions = mdsuite.utils.meta_functions.get_dimensionality(self.box_array)
+        self.volume = np.prod(self.box_array)
+        # todo look into replacing these properties
+        self.sample_rate = int(round(metadata.sample_step / self.time_step))
+        species_list = copy.deepcopy(metadata.species_list)
+        if update_with_pubchempy:
+            update_species_attributes_with_pubchempy(species_list)
+        # store the species information in dict format
+        species_dict = dict()
+        for sp_info in species_list:
+            species_dict[sp_info.name] = {'mass': sp_info.mass,
+                                          'charge': sp_info.charge,
+                                          'properties': [prop_info.name for prop_info in sp_info.properties]}
+        self.species = species_dict
+        # assume the same property for each species
+        self.property_groups = next(iter(species_dict.values()))['properties']
+
+
+def update_species_attributes_with_pubchempy(species_list):
+    """
+    Add information to the species dictionary
+
+    A fundamental part of this package is species specific analysis. Therefore, the
+    Pubchempy package is used to add important species specific information to the
+    class. This will include the charge of the ions which will be used in
+    conductivity calculations.
+
+    """
+    with importlib.resources.open_text(
+            "mdsuite.data", "PubChemElements_all.json"
+    ) as json_file:
+        pse = json.loads(json_file.read())
+
+    # Try to get the species tensor_values from the Periodic System of Elements file
+
+    # set/find masses
+    for sp_info in species_list:
+        for entry in pse:
+            if pse[entry][1] == sp_info.name:
+                sp_info.mass = [float(pse[entry][3])]
+
+                # If gathering the tensor_values from the PSE file was not successful
+    # try to get it from Pubchem via pubchempy
+    for sp_info in species_list:
+        if sp_info.mass is None:
+            try:
+                temp = pcp.get_compounds(sp_info.name, "name")
+                temp[0].to_dict(
+                    properties=[
+                        "atoms",
+                        "bonds",
+                        "exact_mass",
+                        "molecular_weight",
+                        "elements",
+                    ]
+                )
+                sp_info.mass = [temp[0].molecular_weight]
+                log.debug(temp[0].exact_mass)
+            except (ElementMassAssignedZero, IndexError):
+                sp_info.mass = 0.0
+                log.warning(f"WARNING element {sp_info.name} has been assigned mass=0.0")
+    return species_list
+
+
+def _load_trajectory_reader(file_format, trajectory_file, sort: bool = False):
+    # TODO replace by proper instantiation
+    try:
+        class_file_io, file_type = dict_file_io[
+            file_format
+        ]  # file type is per atoms or flux.
+    except KeyError:
+        raise ValueError(
+            f"{file_format} not found! \n"
+            f"Available io formats are are: {[key for key in dict_file_io.keys()]}"
+        )
+    return class_file_io(file_path=trajectory_file, sort=sort), file_type
+
+
+def _species_list_to_architecture_dict(species_list, n_configurations):
+    # TODO let the database handler use the species list directly instead of the dict
+    architecture = {}
+    for sp_info in species_list:
+        architecture[sp_info.name] = {}
+        for prop_info in sp_info.properties:
+            architecture[sp_info.name][prop_info.name] = (
+                sp_info.n_particles,
+                n_configurations,
+                prop_info.n_dim)
+    return architecture
